Use Caffeine for auth caching

+++ enterprise/security/src/main/java/org/neo4j/server/security/enterprise/auth/SecuritySettings.java
@@ -28,7 +28,7 @@ import org.neo4j.helpers.HostnamePort;
 import static org.neo4j.kernel.configuration.Settings.BOOLEAN;
 import static org.neo4j.kernel.configuration.Settings.DURATION;
 import static org.neo4j.kernel.configuration.Settings.HOSTNAME_PORT;
-import static org.neo4j.kernel.configuration.Settings.LONG;
+import static org.neo4j.kernel.configuration.Settings.INTEGER;
 import static org.neo4j.kernel.configuration.Settings.NO_DEFAULT;
 import static org.neo4j.kernel.configuration.Settings.STRING;
 import static org.neo4j.kernel.configuration.Settings.STRING_LIST;
@@ -149,6 +149,6 @@ public class SecuritySettings
             setting( "dbms.security.realms.auth_cache_ttl", DURATION, "10m" );
 
     @Description( "The maximum capacity for authentication and authorization caches (respectively)." )
-    public static Setting<Long> auth_cache_max_capacity =
-            setting( "dbms.security.realms.auth_cache_max_capacity", LONG, "10000" );
+    public static Setting<Integer> auth_cache_max_capacity =
+            setting( "dbms.security.realms.auth_cache_max_capacity", INTEGER, "10000" );
 }
